requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg[binary]
pillow
qrcode
openpyxl
